Bug#18144349 INNODB CANNOT USE THE DOUBLEWRITE BUFFER FOR THE FIRST PAGE OF SYSTEM TABLESPACE
Follow up patch to fix pb2 failures.  During recovery, if a data page is filled
with zeros, and its valid copy is available in the double write buffer, then we
recover the page from the double write buffer.  When this happens a warning was
issued.  This warning is causing many crash recovery tests to fail.  Also, the
usefulness of this warning has not been established.  So warning is removed.

approved by Sunny via IM.