Fix for bug#18235177 DEBUG BUILDS MAY FAIL IN NDBASSERT(FINDTRANSACTION() == ZNOT_FOUND)
Dirty operations (non-transactional) are not inserted into the transactional
hash-list. When deleting such a transaction, there is an assert
'ndbassert(findTransaction() == ZNOT_FOUND) which occasionally fails
      
This assert is too strict, as transactional uniqueness is only required 
for non-dirty operations. tcOpRec ids of dirty operations are reused
within TC (which assign them), and uniqueness for such operations
are not guaranteed, nor required.