WL#6120
-post push fix

After WL#7205 was pushed, mtr tests are also run in
gtid-mode with auto_position=1. The chaged tests here
had CHANGE MASTER TO a file/position related field which
failed saying this can not be done with auto_positio=1.

Fixed by saving the auto_position value in the
beginning of the test, setting auto_position= 0
explicitly before using file/position related options
in the test and restoring auto_position value at the
end of the test.