WL#6885: Flag to indicate session state
In order to ensure session migration from one server to another we need a 
means to know if there is any session context available so that the context 
can be migrated to the target server. This worklog will implement a means to
let the connectors/fabric know that there is a change in session state by 
sending a tracker(flag) to the client in the OK packet.
A session state is considered to change in following cases:
     - user defined variable assignment
     - successful session specific value for server configuration 
       variables assignment
     - temporary tables created, altered or deleted
     - prepared statements added or removed
     - change of the current database