Bug#17786581:SERVER WRITES INTO WORLD-WRITABLE FILE EVEN AFTER WARNING ABOUT IT.
Problem:
========
UUID exists in a file named "auto.cnf", make this auto.cnf
file as world writable. The server starts, and it sees the
auto.cnf is world writable and writes a following warning:
Warning: World-writable con-fig file './auto.cnf' is
ignored

A new UUID is generated and the new UUID is once again
written to the same world writable file which is not safe.

Analysis:
========
At the time of starting a fresh server "auto.cnf" file is
read for getting UUID. Before reading the file first its
permissions are checked if they are valid or not. If the
file is found to be world writable its contents are not
read they are ignored a new UUID is generated as if there
was no previous UUID. But the new UUID was again written to
this world writable file which is incorrect.

Fix:
===
Since the world writable file's contents are ignored there
is no necessity to keep this file. The file is deleted.
A new file with appropriate permissions is created during
new UUID generation.