Bug#18313717: ENABLE -WERROR IN MAINTANER MODE WHEN COMPILING WITH CLANG
Before this patch, the MYSQL_MAINTAINER_MODE CMake option was turned on by
default for debug builds and off for release builds. And MYSQL_MAINTAINER_MODE
caused -Werror to be enabled when building with GCC.

The problem with this was that it was cumbersome to enable -Werror when compiling
with e.g. Clang. This patch fixes the problem by making MYSQL_MAINTAINER_MODE
default on when compiling debug builds with GCC and changing MYSQL_MAINTAINER_MODE
so that it enbles -Werror regardless of if GCC or Clang is used.

This means that enabling -Werror with Clang can simply be done by explicitly
setting -DMYSQL_MAINTAINER_MODE=1 when running CMake.

The patch also fixes a few compilation warnings reported by Clang 3.4, making
it possible to build default debug build with -Werror.