Bug#18313717: ENABLE -WERROR IN MAINTANER MODE WHEN COMPILING WITH CLANG
Post-push fix: Setting -Werror also for "Debug" CMake build type (and not just WITH_DEBUG),
caused a compilation warning in the HAVE_PERRCREED auth plugin CMake check 
to be upgraded to error, causing the check to fail for debug builds. This meant that
auth_socket was not built, leading to RPM build failure on EL6.

This patch fixes the compilation warning, causing the check to pass again.