Bug #18337036: GTID RETRIEVED RECONSTRUCTION SHOULD NOT DEPEND ON THREAD CONTEXT.
Currently, on the relay log initialization, thread context is used as
a flag for the reconstruction of the GTID retrieved set, an operation
that does not depend on this parameter. This is problematic if the
relay log initialization is called in another context different from
the legacy replication scenario.
If the invocation is made in a context where the thread context is
always present this will prevent the reconstruction of this set. The
opposite can also happen leading to a double initialization.

To avoid this, the thread context test is replaced by a new flag that
allows the reconstruction in all contexts but prevents multiple
invocations.