Bug #18353271: HAVE_IB_LINUX_FUTEX CHECK FAILS ON DEBUG BUILD DUE TO MISSING #INCLUDE
The HAVE_IB_LINUX_FUTEX CMake check failed with -Werror 
(which is default on for GCC and debug builds) with:
  error: implicit declaration of function ?syscall?
  [-Werror=implicit-function-declaration]
    return(syscall(SYS_futex, futex, FUTEX_WAIT_PRIVATE, v, NULL, NULL, 0));
    ^
as syscall is defined in unistd.h and this header was not included in the check.

Without -Werror (e.g release builds) the check passed
as syscall was found during linking.

This patch fixes the issue by adding the missing #include statement.

The CMake check was added in WL#6044.