BUG 18225489 - SCHED_GETCPU() CALLS FROM THE FUZZY COUNTER CAUSE PAGE_FAULTS ON VALE29
This affects range reads most because the ached_getcpu() call is used by the
fuzzy counter to count the number of rows read.

For this particular counter we already have a good indexer from
the_get_thread_id() which we can retrieve from the THD instance.

rb#4663, approved by sunny.