Fix for Bug#18345786 CRASH AROUND ST_JOIN_TABLE::AND_WITH_CONDITION. Regression introduced by WL#7082. The necessary ingredients to crash are: SET, ROW (), outer join.
When we are optimizing the most inner subquery, we crash in pushdown_on_conditions (), because outer join information is incorrect, because simplify_joins () was not called. Now let's explain why it was not called.
The statement is SET, this is important, because the first thing it does is locking tables (plain SELECT doesn't, rather does post-prepare locking in mysql_prepare_and_optimize_select()).
Then, we go into fix_fields of this Item_row:
row((select 1 from `t1` left join `t1` `t2` on 1),1)
first it does fix_fields on the subquery argument, which does select_Lex:::prepare () on this subquery (the most inner one, with the left join); during this prepare (), we SKIP apply_local_transforms () (and thus we skip simplify_joins ()); we skip it because we say "I belong to an outer subquery, let it transform me at the end of its own prepare (); transformations are to be done top-down".
Then Item_row:: fix_fields wants to know if the subquery argument is constant, so it calls Item_subselect::const_item (), which says "yes" because tables ARE locked.
Then Item_row:: fix_fields, seeing that the subquery argument is constant, wants to know if it is NULL, thus evaluates it; so we go into JOIN:: optimize for the most inner subquery, hence the crash as simplify_joins () has not been called yet for this subquery.

Before the worklog, there was no problem, because JOIN:: optimize called simplify_joins(). 
Fix: add some robustness: in JOIN::optimize, if simplify_joins() has not been called, call it. This
is expected to be an uncommon case; so far we know only of "set x=(subq);".