WL#7332: Cluster consistent view-id and node attributes propagation
Post fix.
Due to some bug on Corosync version 2.3, the local node id cannot
always be extracted at the first try.
Due to this, sometimes the local node id would be zero causing the
biding to crash.
To solve this problem for now, a loop is introduced that tries more
than once to extract this info from corosync.