Bug #18375920 	NDB TESTING : SYNC_SLAVE_WITH_MASTER CAN TIMEOUT AND CONTINUE
This fix changes mysqltest's sync_slave_with_master 
ndb-binlog-injector-sync timeout from 30s to 150s, and
makes timeout a hard test failure.

This causes the 'fishy' ndb_rpl_commit_afterflush testcase to fail, as
it relied on a timeout+pass behaviour from sync_slave_with_master.

This testcase is removed and a new testcase is added to replace it:
ndb_binlog_flush_tables_with_read_lock.test.

The new testcase verifies the basics of the behaviour, with 2 MySQLDs
attached to a cluster.