BUG#17477959 - VIEW COLUMNS EXCEED TABLE LIMITS
Analysis:
--------

Views cannot be restored from a dump file when the
column names and the number of columns of the view
are very large.

'mysqldump' utility creates tables for each of the
view defined inorder to resolve the interdependencies   
between views. These tables are dropped later when the
actual view is created. When a table is created for a
view having large column names and number of columns,
it hits the limitation of 64KB imposed by the '.frm' file
structure for the table. Hence fails with an error
'Too many columns'.

Fix:
----
Instead of creating tables which are dropped later, create
views which are dropped later before the actual view is
created.