Bug#18137535 - SERVER CRASH ON CREATE TABLE... SELECT
Analysis:
---------
While creating a table having SELECT, first table from the
table global and local table list is unlinked. Later we
check whether we have any table which is opened for update.
If yes then we return from the function by reporting an
error but without linking back the unlinked table to table
lists.
This works well when the LEX object is not cached. But with
prepared statements and stored programs where LEX object of
the statements is cached. Not linking back the first table
would result in unexpected behaviour on the second execution
of the prepared statement or stored programs. 

In the scenario mentioned in the bug report, the local list
became empty on unlinking the first table on first execution
of stored procedure. The assert to check first_table is
not empty while creating the table fails on next execution
of the stored procedure.

Fix:
----
To resolve this, linking back the unlinked first table to
table lists before reporting an error.

RB:
---
http://rb.no.oracle.com/rb/r/4880/