Bug#18383840: SEGV IN MEMCPY(), HA_PARTITION::POSITION
post-push fix for:
Regression from bug#18167648: WRONG RESULTS WITH PARTITIONING,
INDEX_MERGE MYISAM AND NO PK INNODB

The priority queue is setup in ha_partition::index_init() and also sets
m_sec_sort_by_rowid. But in case the index scan is only done on one
partition, due to pruning, the priority queue will not be used and
the ref will not be copied into the priority queue. So only checking
m_sec_sort_by_rowid is not enough to ensure that the ref is
accessible.

Fixed by checking that ordered scan is used.