Bug#18395059: EXPLAIN FOR CONNECTION: DYNAMIC RANGE ACCESS               SHOWS ACCESS TYPE "RANGE"
EXPLAIN FOR CONNECTION showed wrong filtering value for 
dynamic range queries. 

The problem was that Explain_join::explain_rows_and_filtered()
relies on tab->position->fanout to be set to whatever the 
access method produces. In the case of table scan, the number is
the number of rows in the table. For ref access it is the
rec_per_key, for range access it is the range estimate
(select->quick->records). However, when dynamic range is used,
tab->position->fanout remains the fanout of a table scan even
though examined_rows is set to the rows estimate of the current
quick select. This gives incorrect values in the "filtered" 
column.