This patch fixes the bug#17864349 - CHANGING TRUNCATE TABLE TO DROP TABLE & CREATE TABLE MAKES TRIGGER.TEST FAIL
The bug is a regression from WL#6030 and relates to handling of
NOT-NULL columns during execution of UPDATE statement when
there is a BEFORE UPDATE trigger and table whose column is being
updated is not a main table in the statement (that is the update
operation is made using temporary table).

The reason for this bug was that during copying a data of fields from
temporary table to a destination table an information about the fact
that a field was set to null temporary wasn't copied. Also in case
when BEFORE-update trigger was executed against a not-main table in the
query a checking of table fields for NOT-NULL wasn't done after trigger
execution.