Bug#17039528	CRASH WHEN ENABLING INNODB_TABLE_MONITOR 		(ASSERTION: IB_TABLE->STAT_INITIALIZED)
PROBLEM
-------

While saving the statistics to the disk there is
an assert to check if statistics have been 
initialized or not,which was failing. The reason 
for this failure was that the variable 
table->stat_initialized was set as FALSE by 
dict_table_print() which is called by background 
server monitor thread. 

void dict_table_print()
{
  if(!table->stat_initialized)
  {
    dict_stats_update_transient(table);
  }	

  /* print the statistics */
    
  table->stat_initialized = FALSE;
}

If the table statistics are not initialized then
dict_table_print() will update the the statistics
using dict_stats_update_transient() which sets 
table->stat_initialized = TRUE,print the statistics
and then set table->stat_initialized = FALSE to 
bring it back to its original state. This is wrong 
because we are doing it without prior knowledge of 
state of stat_initialized variable (It might be 
TRUE from the beginning). Since the the variable
table->stat_initialized is not a protected member,
it can be set TRUE at any time.

FIX
---
 
We do not need to set table->stat_initialized = FALSE.
The print function just need to print the statistics
and call dict_stats_update_transient() only when 
table->stat_initialized is FALSE. 

[Aprooved by Vasil #rb4944]