Bug #18196562: NDBAPI : RACE BETWEEN LATE TRANSID_AI AND API CLOSETRANSACTION()
An NDBAPI application sends a scan query to a data node, which is 
processed by the TC. The TC forwards the LQHKEYREQ request to the 
appropriate LDM, and aborts the transaction if it does not receive 
a LQHKEYCONF response within the specified time limit. After the 
transaction is successfully aborted, the TC sends a TCROLLBACKREP 
to the NDBAPI client. The NDBAPI client processes the TCROLLBACKREP 
by cleaning up the Ndb objects associated with the transaction.

The NDBAPI client receives the data which it has requested in the 
form of TRANSID_AI signals sent by the LDM. These TRANSID_AI signals 
are buffered for sending at the data node, and may be delivered to the 
NDBAPI client after a delay. On receiving a TRANSID_AI signal, the 
NDBAPI checks the transaction state and ID: if these are as expected, 
it processes the signal using the Ndb objects associated with that 
transaction.

The bug occurs when all the following conditions are fulfilled:

- The TC aborts a transaction due to delays and sends a TCROLLBACPREP 
to the NDBAPI client. Simultaneously, a TRANSID_AI which has been 
buffered for delivery at an LDM is delivered to the same NDBAPI client.

- The NDBAPI client considers the transaction complete on receiving a 
TCROLLBACKREP and immediately closes the transaction.

- The NDBAPI client has a separate receiver thread running concurrently 
with the transaction closing thread.

- The reception of the late TRANSID_AI interleaves with the user 
thread's transaction close so that the TRANSID_AI processing passes the 
normal checks (TransId, 'magic number') before the closeTransaction 
resets the transaction state and invalidates the receiver.

The receiver thread proceeds to continue working on the TRANSID_AI 
signal using the invalidated receiver. Since the receiver is already 
invalidated, its usage results in a core.

Added a fix in the NDBAPI, so that the Ndb object cleanup done for 
TCROLLBACKREP now includes invalidation of the transaction ID. So, for 
a given transaction, any signal which is received after the 
TCROLLBACKREP arrives, does not pass the transaction ID check and is 
silently dropped. Duplicated the fix for TC_COMMITREF, TCROLLBACKREF, 
TCKEY_FAILCONF and TCKEY_FAILREF.