WL#6813: Intra-schema MTS: ordered commits (sequential consistency)
DESCRIPTION
===========
This worklog ensures that the commits by slave worker threads running in
parallel are done in the same order as on the master. This also means that
the slave will never externalize a database state which was never externalized
by the master.

DESIGN
======
- slave_preserve_commit_order
  The behavior is optional. Users can control it through a new system variable
  named slave_preserve_commit_order.

- Commit_order_manager
  We implement a commit order manager which maintains a queue of transactions
  which are applying by workers.

  Coordinator registers them into the queue when dispatching them to a worker.
  So the transactions in the queue have the order as they are in the relay log.

  All the transactions should go into the flush stage(see Binlog Group commit)
  in the queue sequence. When a worker commits its transaction(before the
  transaction goes into flush stage), it must wait until it becomes the queue
  head. It does means all transactions before it have entered into flush stage.

  Once the transaction enters flush stage, It is popped from the queue
  and signals the next transaction to go ahead.

The feature is for only intra-schema multi-threaded slave, but not database
partitioned multi-threaded slave. And it is not supported when log_bin or
log_slave_updates is OFF.