WL#6819: GCS Replication: Overview
Post-fix for several Valgrind reported errors. 

The solved errors:

1)
In the Certification handler a pipeline_event could be sometimes
deleted and replaced by another event. The problem is that, when the
pipeline returns, the event pointer goes back the original value. i.e,
pointing to a already freed memory block. This causes a double free
invocation error. To avoid this, the event is no longer deleted but
reused.

2) 
The creation of the generic function for registration of
PSI_mutex_info and PSI_cond_info arrays in the gcs_plugin_utils file
caused a bug that avoid the correct execution of this action. 
In fact, the number of elements on these arrays was determined by
array_elements(), a method based on the memory sizes of the whole
array and it's first element. In a generic method such as this one
that receives arrays, the size of an array is always 8 (the pointer
size) and so the calculation of it's elements is always wrong. 
To avoid this, the registration method now receives the arrays' sizes

3) 
In the Continuation class, the class lock and condition PSI_keys were not
registered causing some warnings about their lack of initialization.
They are now registered.

4)
There were warnings about the Continuations transaction_discarded flag
non initialization. The flag is now initialized.

5) 
The data written value used in the event to packet conversion is
sometimes uninitialized (when we inject GTIDs, for example) so it's
value was sometimes set to garbage. The value is now always
initialized to 0. 
Related to this, this bug showed that we sometimes pass 0 as the size
to the cache used on these conversions. As the cache when given a 0
value replaces it by a default value, this doesn't seems to be an
issue.

6)
On the add_pke_to_list method, when using the my_safe_itoa method, the
last pointer value passed to the function is wrong as it uses the
sizeof function. This returns the size of the pointer (8) and not the
array size. We now use the correct length for this operation.

7) 
Replaced some deletes for deletes[] one some situations where the
first was wrongly used.

Memory leaks:

1) 
On the get_pipeline method the handler_list is created but never
released.

2)
Also on the pipeline creation, on the configure_pipeline method, if
there is an error with the handler list, the last handler is not
deleted.

3) 
In the Pipeline event class, on deletion neither packet nor log event
are deleted.

4)
On the Certification handler the Certifier is created but never
deleted.

5) 
On the Applier loop, neither continuations or format description
events are deleted.

6)
In the Transaction context event there were some issues with my_free
method and my_strndup functions that resulted in some leak warnings.
These are now solved with a simple cast.