Bug#18363910 SKIP MAIN.PLUGIN_LOAD AND MAIN.PLUGIN_LOAD_OPTIONS ON WINDOWS EMBEDDED
The tests plugin_load and plugin_load_options will pre-load the plugin
*before* the test is started, so the 'skip this on windows' logic is executed too late.

Plugin initialization fails because the embedded server is compiled without
performance schema, and the plugin is compiled with performance schema.
In this case, mysql_mutex_register has two different implementations,
breaking the "one-definition-rule". The trick of mapping PSI_server to
PSI_noop works on most platforms, but not windows.

Solution: remove the mutex registration code, the mutex was unused anyways.