Bug #18271429 - GB18030_UNICODE_520_CI CRASHES PARSER!
The gb18030_unicode_520_ci charset is not set with a proper ctype,
so the state_map of it would not be initialized. When parser want
to use the state_map, server crashes.

We should give a proper ctype which shall be the same with that
defined in gb18030.c to gb18030_unicode_520_ci charset.

rb://4732, approved by Ramil.