Bug#18436558 ALTER TABLE ADD COLUMN FAIL TO COPY HASHMAP TO NEW TABLE DEFINITION
For ALTER TABLE there table schema are changed without change of
partitioning, the new table definition will not copy the hashmap from
the old defintion but use the current default hashmap - but will not
reorganize data according to the new hashmap!

This makes some rows inaccessible via primary key lookup if the current
default hashmap is not compatible with the hashmap of the old definition.

Checks are added both in kernel and in ndbapi that an alter table request will not
contain a hashmap change without also trigger reorganisation.

Also the hashmap will now be copied when copying a new table definition from an old.

Test ndb_addnode will fail if one do not apply the part of patch that changes
NdbTableImpl::assign().