Bug #18103035: WRONG BISON LOCATION FOR EMPTY RULE RIGHT HAND SIDE ==================================================================
The bug in in the custom definition of the YYLLOC_DEFAULT macro:
for the empty right hand sides (RHS) of the grammar rules it
returns exactly same "location" data as for the previously parsed
token.

I.e. for the sample grammar:

    empty_rule_example: /*empty*/;

    some_rule_example: SOME_TOKEN empty_rule_example
    {
      printf("%s, %s\n", @1.start, @1.end);
      printf("%s, %s\n", @2.start, @2.end);
    }
    ;

both printf() function calls output the same data.
Obviously this is not good that the empty non-terminal is
associated with a non-empty location data.

The fix:

The solution is to return the end of previously parsed token
as the start (and the end as before) of the empty RHS.

In this case @$.start == @$.end for the empty token and the
calling rule doesn't refer to unexpected locations of previously
parsed tokens from the outer rules.

The known minor problem of the fix:

As far as tokens may be delimited with whitespaces, the grammar:

    empty_rule: /* empty */;

    outer_rule: "OUTER_TOKEN" inner_rule;

    inner_rule: empty_rule "TAIL_TOKEN";
    {
       printf("%s\n", @$.start);
    }
    ;

and the input of "OUTER_TOKEN TAIL_TOKEN" will output a bit
confusing " TAIL_TOKEN" string with leading space instead of
simple "TAIL_TOKEN" string.
This happens because the start of the empty non-terminal
location is the end of the previously parsed token.

To omit unwanted whitespace delimiters we can apply a
workaround where this is necessary:

    empty_rule: /* empty */;

    rule_with_empty_heading: empty_rule SOME_TOKEN_2 SOME_TOKEN_3
    {
      @$.start= @2.start; // override implicit YYLLOC_DEFAULT
      @$.raw_start= @2.raw_start; // override YYLLOC_DEFAULT
      ...
    }
    ;

    // or for trivial cases:
    trivial_rule_with_empty_heading: empty_rule SOME_TOKEN_2
    {
      @$= @2;
      ...
    }
    ;