WL#7142: Completely ignore *.isl files during redo log apply. Instead, use the *.ibd file names from MLOG_FILE_NAME records.
fil_load_single_table_tablespace(): Replace df_default and df_remote
with Datafile file. Pass the ownership of the allocated name to
Datafile, so that the ut_free() cleanup will be eliminated, done as
part of Datafile::shutdown(), executed by the destructor. Likewise,
remove calls to Datafile::close(), because that will be part of
shutdown().

Datafile::init(): Add a variant that takes ownership of the "name",
and allows filepath to be initialized.

Datafile::shutdown(): Remove a redundant check for m_name!=NULL.
free(NULL) is documented as no-op in the C standard.