Bug#18413646: SINCE 5.7 ERROR 1416 IS CONSIDERED FATAL, BUT IT SHOULDN'T BE
The problem was that the handler error HA_ERR_NULL_IN_SPATIAL was 
considered fatal, which mean it was not possible to write a SP HANDLER
which would trigger if this error was reported.

This was a regression introduced by the patch for Bug#16587369
which changed how handler errors are determined to be fatal
(or ignorable).

This patch fixes the problem by adding HA_ERR_NULL_IN_SPATIAL
to the list of non-fatal handler errors.