Bug#18475390: STOP SLAVE TIMEOUT IS SET STATICALLY AT EACH SLAVE THREAD TERMINATION
Whenever the terminate method is called for the slave SQL/IO threads,
inside this, a timeout variable associated to this task is always set
based on a server variable.
The problem with such routine is that we cannot reuse this code to be
used with other defined timeouts.

As the timeout can be passed as a parameter from method to method,
instead of being statically set and used as of now, the code is
changed to implement this pattern allowing the method reuse in
different contexts.