Bug#18364815: OPTIMIZER PREFERS TABLE SCAN WHEN USING "IN"               WITH VALUE OF DIFFERENT TYPE
Item_func_in checked whether there were items of different type
in the IN clause and would automatically reject the use of 
range access in that case. However, this is too restrictive 
since a great deal of type conversions are automatically done
in MySQL, e.g. conversion of quoted numbers to integer.
"col IN (1, "2")" is therefore perfectly legal but used to be
rejected.
      
The fix is to remove the automatic rejection and let the range 
optimizer make the decision on its own since it already 
performs this check on its own (see comparable_in_index())
anyway.