BUG#18482854 RPL : ROTATE_LOG_EVENT INCORRECTLY ADVANCES                    GROUP_RELAY_LOG_POS IN A GROUP
Problem:
=======

Rotate events can cause the group_relay_log_pos to be incorrectly moved
forward within a group.

This means that when the transaction is retried, or if you stop the SQL
thread in the middle of a transaction after some Rotates (considering
that the transaction/group was spanned into multiple relay log files),
the group (or part of the group from the beginning) will be silently
skipped.

Analysis:
========

It was found a problem in the logic for avoiding to touch SQL thread
coordinates in Rotate_log_event::do_update_pos().

The logic was allowing to update SQL thread coordinates if not using
MTS regardless rli->is_in_group() signaling that it was in the middle
of a group.

Fix:
===

Corrected the logic.