Bug#18483776: ASSERTION FAILED: QUERY_SOURCE == 0,               FILE SQL_PROFILE.CC, LINE 297
This assert was triggered since PROFILING::set_query_source() by
mistake was called twice for a prepared statement. On release
builds this could lead to memory leaks.

This bug was a regression introduced by the patch for Bug#17606098
which moved the setting of the profiling query string from
mysqld_stmt_execute() to Prepared_statement::execute_loop().
This patch fixes the problem by moving the call back.