wl#7332 (post-push) GCS: missing handle_view_change() call and stats update when group loses quorum
After a view change that makes a cluster member to depart the quorum
group, the member still displays the former quorum configuration
through P_S.

The reason of that was that non-quorate view-changes were not made
to call Protocol::"view_change" the user-lever handler.

Fixed with letting this invocation.

Notice I do not touch in this commit an existing rule of assigning 0
as a value to non-quorate group members.
Preserving the former quorum group view_id is deem of a little value,
at the same time the view_id of 0 is explicit about the actual status
of such group and its members.